Un-indent project relation, reset tac crunchbase
When a project is listed under 'hosted' in the relations settings, it's
expected they will not have their own crunchbase, else the build will
complain and fail. This is fixed by resetting LFN tac projects to
lf-networking crunchbase.

This also modifies the relation list to unindent non-lfn projects from
being children of the hosted category.

Signed-off-by: Trevor Bramwell <tbramwell@linuxfoundation.org>